changed to use only first 6 characters of title
there was something that started happening where the extensions were
added to the part titles which were previously only the name, and now
they had the extension too. This broke a bunch of stuff, so i'm now only
asking for the first 6 characters of the name. This could probably be
done a lot more cleanly...